+++ cpp/api/include/Object.h
@@ -18,6 +18,7 @@
 #define CAMEO_OBJECT_H_
 
 #include <string>
+#include <atomic>
 
 namespace cameo {
 
@@ -27,6 +28,11 @@ namespace cameo {
 class Object {
 
 public:
+	/**
+	 * Constructor.
+	 */
+	Object();
+
 	/**
 	 * Destructor.
 	 */
@@ -38,15 +44,43 @@ public:
 	virtual void init() = 0;
 
 	/**
+	 * Returns true if is ready.
+	 * \return True if is ready.
+	 */
+	bool isReady() const;
+
+	/**
 	 * Terminates the object.
 	 */
 	virtual void terminate() = 0;
+
+	/**
+	 * Returns true if is terminated.
+	 * \return True if is terminated.
+	 */
+	bool isTerminated() const;
 
 	/**
 	 * Returns a string representation of this application.
 	 * \return The string representation.
 	 */
 	virtual std::string toString() const = 0;
+
+protected:
+	/**
+	 * Sets the state ready.
+	 */
+	void setReady();
+
+	/**
+	 * Sets the state terminated.
+	 */
+	void setTerminated();
+
+private:
+	enum class InternalState {INIT, READY, TERMINATED};
+
+	std::atomic<InternalState> m_state = InternalState::INIT;
 };
 
 }