+++ java/api/src/main/java/fr/ill/ics/cameo/coms/Subscriber.java
@@ -21,10 +21,9 @@ import org.json.simple.JSONObject;
 import fr.ill.ics.cameo.base.App;
 import fr.ill.ics.cameo.base.App.Com.KeyValueGetter;
 import fr.ill.ics.cameo.base.ICancelable;
-import fr.ill.ics.cameo.base.IObject;
 import fr.ill.ics.cameo.base.ITimeoutable;
 import fr.ill.ics.cameo.base.InitException;
-import fr.ill.ics.cameo.base.KeyValueGetterException;
+import fr.ill.ics.cameo.base.StateObject;
 import fr.ill.ics.cameo.base.This;
 import fr.ill.ics.cameo.base.Timeout;
 import fr.ill.ics.cameo.base.TimeoutCounter;
@@ -40,7 +39,7 @@ import fr.ill.ics.cameo.strings.StringId;
 /**
  * Class defining a subscriber.
  */
-public class Subscriber implements IObject, ITimeoutable, ICancelable {
+public class Subscriber extends StateObject implements ITimeoutable, ICancelable {
 	
 	private App app;
 	private String publisherName;
@@ -142,6 +141,8 @@ public class Subscriber implements IObject, ITimeoutable, ICancelable {
 		catch (Exception e) {
 			throw new InitException("Cannot create subscriber: " + e.getMessage());
 		}
+		
+		setReady();
 	}
 	
 	/**
@@ -240,8 +241,11 @@ public class Subscriber implements IObject, ITimeoutable, ICancelable {
 	 */
 	@Override
 	public void terminate() {
+		
 		waiting.remove();
 		impl.terminate();
+		
+		setTerminated();
 	}
 
 	@Override