+++ tests/cpp/src/TestSimple.cpp
@@ -44,7 +44,10 @@ int main(int argc, char *argv[]) {
 	}
 
 	unique_ptr<Server> server = Server::create(endpoint, useProxy);
+
+	cout << "Server ready ? " << server->isReady() << endl;
 	server->init();
+	cout << "Server ready ? " << server->isReady() << endl;
 
 	// Loop the number of times.
 	for (int i = 0; i < numberOfTimes; ++i) {