+++ tests/cpp/src/TestSubscriber.cpp
@@ -72,7 +72,10 @@ int main(int argc, char *argv[]) {
 
 		// The publisher is created after the applications that will wait for it.
 		unique_ptr<coms::Publisher> publisher = coms::Publisher::create("publisher");
+
+		cout << "Publisher ready ? " << publisher->isReady() << endl;
 		publisher->init();
+		cout << "Publisher ready ? " << publisher->isReady() << endl;
 
 		for (int k = 0; k < 20; ++k) {