+++ java/api/CHANGELOG.md
@@ -2,6 +2,7 @@
 -----
 
 * Requester and subscriber init() can timeout.
+* Added isReady() and isTerminated() to parent class of coms classes.
 
 2.1.0
 -----